Rename components: Setting→SettingItem, Footer→LayoutFooter, Header→LayoutHeader, RouterTitle→LayoutTitle, UpdateHandlerDialog→UpdateDialog, Bookmark→BookmarkItem
